Audyt stabilnosci: naprawa 13 potwierdzonych bledow (pamiec/OOM, utrata danych, crash na danych)
Po naprawie crash-petli - adwersaryjny audyt 5 wymiarow (47 agentow) znalazl i potwierdzil
kolejne bledy tej samej klasy. Naprawione:

UTRATA DANYCH / SYNC (lamalo zasade "nic nie moze zginac"):
- offloadSkany: zapisywal NIEAKTUALNY snapshot skanu -> gubil rownolegle edycje i WSKRZESZAL
  usuniete skany. Teraz: uploady, potem read-modify-write na AKTUALNYM rekordzie (find po id),
  pomijamy usuniety (uszanowany tombstone), podmieniamy TYLKO realnie wgrane strony.
  Zweryfikowane 4 niezmienniki (edycja przetrwa / brak wskrzeszenia / nowa strona zostaje / brak zbednego zapisu).
- wgrajSkanStrone: polykal blad braku bucketa 'skany' -> offload cicho nie dzialal w kolko.
  Teraz rozroznia brak-bucketa od sieci: pokazuje JEDEN czytelny komunikat (uruchom amico-skany.sql)
  i przestaje bic co 45 s; flaga kasowana przy wznowieniu/reconnekcie (po SQL wraca sam).
- Limit 20 MB: komunikat nie zwala juz wszystkiego na skany; log pokazuje rozklad rozmiaru
  (skany vs podpisy/logo vs metadane), zeby diagnozowac realne zrodlo.
- Tombstones: retencja 90 dni -> 5 lat (znikomy rozmiar, a nie wskrzesza usunietych rekordow
  z urzadzenia dlugo offline).

CRASH NA DANYCH (bialy ekran):
- Pulpit/Klienci/Wyceny/Faktury/Skany/Umowy/... : ~15 sortowan .localeCompare na polach
  utworzono/zaktualizowano/data ktore moga byc undefined (stary rekord / import) -> teraz (x||'').
- wycenaStatusInfo/umowaStatusInfo zwracaly undefined dla nieznanego/pustego statusu -> .tone
  kladlo caly widok. Dodany fallback. (Znalezione testem: /umowy padalo na umowie bez statusu.)

PAMIEC / OBRAZY (iPhone PWA):
- SkanImg: domyslnie loading="lazy" + decoding="async" -> siatki/podglady nie dekoduja
  wszystkich duzych skanow naraz (glowna przyczyna ubijania karty na archiwum/wielostronicowych).
- scanner adaptiveThreshold: Float64Array -> Int32Array (dokladny, polowa pamieci: ~14 zamiast ~27 MB).
- store.persist: debounce 250->700 ms + pomijanie zapisu, gdy referencja bazy sie nie zmienila.
- zastapBaze: kasuje zaplanowany persist i nie dubluje zapisu tej samej bazy.

SYNC - EFEKTYWNOSC:
- Wznowienie PWA odpalalo 2-4 pelne pobrania bazy (visibilitychange+focus+SUBSCRIBED). Teraz
  koalescencja (jedna obietnica) + tanie sprawdzenie rev (pelna baza tylko gdy serwer nowszy).

Weryfikacja: tsc+build czyste; harness 23 tras z rekordami bez status/dat/tablic = 0 crashy;
audyt mobilny 0 problemow; flows OK; migracja idempotentna (bajt-w-bajt).

+++ src/lib/scanner.ts
@@ -203,7 +203,10 @@ function adaptiveThreshold(
   C: number,
 ): Uint8ClampedArray {
   const W = w + 1
-  const integ = new Float64Array(W * (h + 1))
+  // Obraz calkowy (integral image). Sumy sa CALKOWITE i <= w*h*255 (dla max ~2200 px to
+  // ~8.7e8 << 2^31), wiec Int32Array jest DOKLADNY i zajmuje polowe pamieci Float64
+  // (~14 MB zamiast ~27 MB) - mniejszy szczyt pamieci przy skanowaniu na iPhone.
+  const integ = new Int32Array(W * (h + 1))
   for (let y = 0; y < h; y++) {
     let rowsum = 0
     for (let x = 0; x < w; x++) {
